requirements.txt
Django==6.0.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
django-cors-headers==4.3.0
django-filter==25.1
psycopg2-binary==2.9.9
